MAINT: antenna to cart. funcs in util.transforms
