+++ doc/source/dev_reference/util.rst
@@ -6,4 +6,5 @@ Miscellaneous utility functions.
 
 .. automodule:: pyart.util.circular_stats
 .. automodule:: pyart.util.hildebrand_sekhon
+.. automodule:: pyart.util.transforms
 .. automodule:: pyart.util.xsect
