+++ pyart/graph/common.py
@@ -8,10 +8,8 @@ Common graphing routines.
     :toctree: generated/
 
     dms_to_d
-    radar_coords_to_cart
     corner_to_point
     ax_radius
-    sweep_coords_to_cart
     parse_ax
     parse_ax_fig
     parse_vmin_vmax
@@ -31,10 +29,6 @@ Common graphing routines.
     generate_vpt_title
     generate_ray_title
     set_limits
-    _interpolate_range_edges
-    _interpolate_elevation_edges
-    _interpolate_azimuth_edges
-    _interpolate_axes_edges
 
 """
 
@@ -42,10 +36,9 @@ import numpy as np
 import matplotlib.pyplot as plt
 from netCDF4 import num2date
 
+from ..io.common import dms_to_d
 
 PI = 3.141592653589793
-
-from ..io.common import dms_to_d, radar_coords_to_cart
 
 
 def corner_to_point(corner, point):
@@ -100,90 +93,6 @@ def ax_radius(lat, units='radians'):
         const = 1.0
     R = Re * np.sin(PI / 2.0 - abs(lat * const))
     return R
-
-
-def sweep_coords_to_cart(ranges, azimuths, elevations, edges=False):
-    """
-    Calculate Cartesian coordinate for the gates in a single radar sweep.
-
-    Calculates the Cartesian coordinates for the gate centers or edges for
-    all gates in a single sweep assuming a standard atmosphere (4/3 Earth's
-    radius model). See :py:func:`pyart.io.common.radar_coords_to_cart` for
-    details.
-
-    Parameters
-    ----------
-    ranges : array, 1D.
-        Distances to the center of the radar gates (bins) in meters.
-    azimuths : array, 1D.
-        Azimuth angles of the sweep in degrees.
-    elevations : array, 1D.
-        Elevation angles of the sweep in degrees.
-    edges : bool, optional
-        True to calculate the coordinates of the gate edges by interpolating
-        between gates and extrapolating at the boundaries.  False to
-        calculate the gate centers.
-
-    Returns
-    -------
-    x, y, z : array, 2D
-        Cartesian coordinates in meters from the center of the radar to the
-        gate centers or edges.
-
-    """
-    if edges:
-        if len(ranges) != 1:
-            ranges = _interpolate_range_edges(ranges)
-        if len(elevations) != 1:
-            elevations = _interpolate_elevation_edges(elevations)
-        if len(azimuths) != 1:
-            azimuths = _interpolate_azimuth_edges(azimuths)
-    rg, azg = np.meshgrid(ranges, azimuths)
-    rg, eleg = np.meshgrid(ranges, elevations)
-    return radar_coords_to_cart(rg / 1000., azg, eleg)
-
-
-def _interpolate_range_edges(ranges):
-    """ Interpolate the edges of the range gates from their centers. """
-    edges = np.empty((ranges.shape[0] + 1, ), dtype=ranges.dtype)
-    edges[1:-1] = (ranges[:-1] + ranges[1:]) / 2.
-    edges[0] = ranges[0] - (ranges[1] - ranges[0]) / 2.
-    edges[-1] = ranges[-1] - (ranges[-2] - ranges[-1]) / 2.
-    edges[edges < 0] = 0    # do not allow range to become negative
-    return edges
-
-
-def _interpolate_elevation_edges(elevations):
-    """ Interpolate the edges of the elevation angles from their centers. """
-    edges = np.empty((elevations.shape[0]+1, ), dtype=elevations.dtype)
-    edges[1:-1] = (elevations[:-1] + elevations[1:]) / 2.
-    edges[0] = elevations[0] - (elevations[1] - elevations[0]) / 2.
-    edges[-1] = elevations[-1] - (elevations[-2] - elevations[-1]) / 2.
-    edges[edges > 180] = 180.   # prevent angles from going below horizon
-    edges[edges < 0] = 0.
-    return edges
-
-
-def _interpolate_azimuth_edges(azimuths):
-    """ Interpolate the edges of the azimuth angles from their centers. """
-    edges = np.empty((azimuths.shape[0]+1, ), dtype=azimuths.dtype)
-    # perform interpolation and extrapolation in complex plane to
-    # account for periodic nature of azimuth angle.
-    az = np.exp(1.j*np.deg2rad(azimuths))
-    edges[1:-1] = np.angle((az[1:] + az[:-1]) / 2., deg=True)
-    edges[0] = np.angle(az[0] - (az[1] - az[0]) / 2., deg=True)
-    edges[-1] = np.angle(az[-1] - (az[-2] - az[-1]) / 2., deg=True)
-    edges[edges < 0] += 360     # range from [-180, 180] to [0, 360]
-    return edges
-
-
-def _interpolate_axes_edges(axes):
-    """ Interpolate the edges of the axes gates from their centers. """
-    edges = np.empty((axes.shape[0] + 1, ), dtype=axes.dtype)
-    edges[1:-1] = (axes[:-1] + axes[1:]) / 2.
-    edges[0] = axes[0] - (axes[1] - axes[0]) / 2.
-    edges[-1] = axes[-1] - (axes[-2] - axes[-1]) / 2.
-    return edges
 
 
 ########################
