+++ pyart/io/common.py
@@ -11,7 +11,6 @@ Input/output routines common to many file formats.
     dms_to_d
     stringarray_to_chararray
     _test_arguments
-    radar_coords_to_cart
     make_time_unit_str
     add_2d_latlon_axis
 
@@ -104,64 +103,6 @@ def _test_arguments(dic):
     if dic:
         import warnings
         warnings.warn('Unexpected arguments: %s' % dic.keys())
-
-
-# XXX move this to another module
-def radar_coords_to_cart(rng, az, ele, debug=False):
-    """
-    Calculate Cartesian coordinate from radar coordinates
-
-    Parameters
-    ----------
-    rng : array
-        Distances to the center of the radar gates (bins) in kilometers.
-    az : array
-        Azimuth angle of the radar in degrees.
-    ele : array
-        Elevation angle of the radar in degrees.
-
-    Returns
-    -------
-    x, y, z : array
-        Cartesian coordinates in meters from the radar.
-
-    Notes
-    -----
-    The calculation for Cartesian coordinate is adapted from equations
-    2.28(b) and 2.28(c) of Doviak and Zrnic [1]_ assuming a
-    standard atmosphere (4/3 Earth's radius model).
-
-    .. math::
-
-        z = \\sqrt{r^2+R^2+r*R*sin(\\theta_e)} - R
-
-        s = R * arcsin(\\frac{r*cos(\\theta_e)}{R+z})
-
-        x = s * sin(\\theta_a)
-
-        y = s * cos(\\theta_a)
-
-    Where r is the distance from the radar to the center of the gate,
-    :math:`\\theta_a` is the azimuth angle, :math:`\\theta_e` is the
-    elevation angle, s is the arc length, and R is the effective radius
-    of the earth, taken to be 4/3 the mean radius of earth (6371 km).
-
-    References
-    ----------
-    .. [1] Doviak and Zrnic, Doppler Radar and Weather Observations, Second
-        Edition, 1993, p. 21.
-
-    """
-    theta_e = ele * np.pi / 180.0       # elevation angle in radians.
-    theta_a = az * np.pi / 180.0        # azimuth angle in radians.
-    R = 6371.0 * 1000.0 * 4.0 / 3.0     # effective radius of earth in meters.
-    r = rng * 1000.0                    # distances to gates in meters.
-
-    z = (r ** 2 + R ** 2 + 2.0 * r * R * np.sin(theta_e)) ** 0.5 - R
-    s = R * np.arcsin(r * np.cos(theta_e) / (R + z))  # arc length in m.
-    x = s * np.sin(theta_a)
-    y = s * np.cos(theta_a)
-    return x, y, z
 
 
 def make_time_unit_str(dtobj):
